# api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/locations/{locationId}": {
      "get": {
        "operationId": "LocationController_findOne",
        "parameters": [
          {
            "name": "locationId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LocationDto"
                }
              }
            }
          },
          "401": {
            "description": "UNAUTHENTICATED: no valid session",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "UNAUTHORIZED: signed in, but not allowed to do this",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "NOT_FOUND: the resource doesn't exist, or isn't visible to you",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "429": {
            "description": "RATE_LIMITED: too many requests, retry after the Retry-After header's seconds",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          }
        },
        "summary": "Get a location's own fields",
        "tags": [
          "Locations"
        ]
      },
      "patch": {
        "operationId": "LocationController_update",
        "parameters": [
          {
            "name": "locationId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateLocationDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LocationDto"
                }
              }
            }
          },
          "400": {
            "description": "VALIDATION_ERROR: the request is invalid",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "UNAUTHENTICATED: no valid session",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "UNAUTHORIZED: signed in, but not allowed to do this",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "NOT_FOUND: the resource doesn't exist, or isn't visible to you",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "429": {
            "description": "RATE_LIMITED: too many requests, retry after the Retry-After header's seconds",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          }
        },
        "summary": "Update a location's settings",
        "tags": [
          "Locations"
        ]
      },
      "delete": {
        "operationId": "LocationController_remove",
        "parameters": [
          {
            "name": "locationId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          },
          "401": {
            "description": "UNAUTHENTICATED: no valid session",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "UNAUTHORIZED: signed in, but not allowed to do this",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "NOT_FOUND: the resource doesn't exist, or isn't visible to you",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "429": {
            "description": "RATE_LIMITED: too many requests, retry after the Retry-After header's seconds",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          }
        },
        "summary": "Delete a location for good, owner only (ADR 0017)",
        "tags": [
          "Locations"
        ]
      }
    },
...
    "/locations/{locationId}/invitations": {
      "post": {
        "operationId": "InvitationController_create",
        "parameters": [
          {
            "name": "locationId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateInvitationDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvitationDto"
                }
              }
            }
          },
          "400": {
            "description": "VALIDATION_ERROR: the request is invalid",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "UNAUTHENTICATED: no valid session",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "UNAUTHORIZED: signed in, but not allowed to do this",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "NOT_FOUND: the resource doesn't exist, or isn't visible to you",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "409": {
            "description": "CONFLICT: not possible in the resource's current state",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "429": {
            "description": "RATE_LIMITED: too many requests, retry after the Retry-After header's seconds",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          }
        },
        "summary": "Invite someone by email to a location with a role (an admin invites employees only)",
        "tags": [
          "Invitations"
        ]
      },
      "get": {
        "operationId": "InvitationController_list",
        "parameters": [
          {
            "name": "locationId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/InvitationDto"
                  }
                }
              }
            }
          },
          "401": {
            "description": "UNAUTHENTICATED: no valid session",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "UNAUTHORIZED: signed in, but not allowed to do this",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "NOT_FOUND: the resource doesn't exist, or isn't visible to you",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "429": {
            "description": "RATE_LIMITED: too many requests, retry after the Retry-After header's seconds",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          }
        },
        "summary": "List invitations for a location",
        "tags": [
          "Invitations"
        ]
      }
    },
    "/locations/{locationId}/invitations/{invitationId}/revoke": {
      "post": {
        "operationId": "InvitationController_revoke",
        "parameters": [
          {
            "name": "locationId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "invitationId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvitationDto"
                }
              }
            }
          },
          "401": {
            "description": "UNAUTHENTICATED: no valid session",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "UNAUTHORIZED: signed in, but not allowed to do this",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "NOT_FOUND: the resource doesn't exist, or isn't visible to you",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "409": {
            "description": "CONFLICT: not possible in the resource's current state",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "429": {
            "description": "RATE_LIMITED: too many requests, retry after the Retry-After header's seconds",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          }
        },
        "summary": "Revoke a pending invitation (an admin revokes employee invitations only)",
        "tags": [
          "Invitations"
        ]
      }
    },
    "/locations/{locationId}/invitations/{invitationId}/resend": {
      "post": {
        "operationId": "InvitationController_resend",
        "parameters": [
          {
            "name": "locationId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "invitationId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvitationDto"
                }
              }
            }
          },
          "401": {
            "description": "UNAUTHENTICATED: no valid session",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "UNAUTHORIZED: signed in, but not allowed to do this",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "NOT_FOUND: the resource doesn't exist, or isn't visible to you",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "409": {
            "description": "CONFLICT: not possible in the resource's current state",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "429": {
            "description": "RATE_LIMITED: too many requests, retry after the Retry-After header's seconds",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          }
        },
        "summary": "Email a pending invitation again, extending its expiry (ADR 0017)",
        "tags": [
          "Invitations"
        ]
      }
    },
...
  },
  "info": {
    "title": "ReadyYet API",
    "description": "Better Auth's own routes (/api/auth/*) aren't included here, they're raw middleware, not Nest controllers. See ADR 0008/0011.",
    "version": "0.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "better-auth.session_token"
      }
    },
    "schemas": {
      "ValidationIssueDto": {
        "type": "object",
        "properties": {
          "constraints": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            }
          },
          "property": {
            "type": "string"
          }
        },
        "required": [
          "property"
        ]
      },
      "ApiErrorDto": {
        "type": "object",
        "properties": {
          "code": {
            "enum": [
              "VALIDATION_ERROR",
              "UNAUTHENTICATED",
              "UNAUTHORIZED",
              "REAUTHENTICATION_REQUIRED",
              "NOT_FOUND",
              "CONFLICT",
              "RATE_LIMITED",
              "INTERNAL_ERROR"
            ],
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "details": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ValidationIssueDto"
            }
          },
          "requestId": {
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ]
      },
      "ApiErrorResponseDto": {
        "type": "object",
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ApiErrorDto"
          }
        },
        "required": [
          "error"
        ]
      },
...
      "LocationDto": {
        "type": "object",
        "properties": {
          "locale": {
            "enum": [
              "EN",
              "FR"
            ],
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "businessId": {
            "type": "string"
          },
          "businessTypeId": {
            "type": "number"
          },
          "name": {
            "type": "string"
          },
          "contactPhone": {
            "type": "string"
          },
          "contactEmail": {
            "type": "string"
          },
          "logoUrl": {
            "type": "string",
            "nullable": true
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          },
          "deletedAt": {
            "format": "date-time",
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "locale",
          "id",
          "businessId",
          "businessTypeId",
          "name",
          "contactPhone",
          "contactEmail",
          "logoUrl",
          "createdAt",
          "updatedAt",
          "deletedAt"
        ]
      },
...
      "UpdateLocationDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "contactPhone": {
            "type": "string"
          },
          "contactEmail": {
            "type": "string",
            "format": "email"
          },
          "logoUrl": {
            "type": "string"
          },
          "locale": {
            "type": "string",
            "enum": [
              "EN",
              "FR"
            ]
          }
        }
      },
...
      "CreateInvitationDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "format": "email"
          },
          "role": {
            "type": "string",
            "enum": [
              "ADMIN",
              "EMPLOYEE"
            ]
          }
        },
        "required": [
          "email",
          "role"
        ]
      },
      "InvitationDto": {
        "type": "object",
        "properties": {
          "role": {
            "enum": [
              "ADMIN",
              "EMPLOYEE"
            ],
            "type": "string"
          },
          "status": {
            "enum": [
              "PENDING",
              "ACCEPTED",
              "REVOKED",
              "EXPIRED"
            ],
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "locationId": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "invitedBy": {
            "type": "string"
          },
          "expiresAt": {
            "format": "date-time",
            "type": "string"
          },
          "acceptedAt": {
            "format": "date-time",
            "type": "string",
            "nullable": true
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "role",
          "status",
          "id",
          "locationId",
          "email",
          "invitedBy",
          "expiresAt",
          "acceptedAt",
          "createdAt"
        ]
      },
...
    }
  }
}