# api/src/invitation/dto/invitation.response.dto.ts
import { ApiProperty } from "@nestjs/swagger";
import { InvitationStatus, Role } from "@readyyet/db";

export class InvitationDto {
  id!: string;
  locationId!: string;
  email!: string;
  // An invitation never makes someone the Owner (ADR 0017).
  @ApiProperty({ enum: [Role.ADMIN, Role.EMPLOYEE] })
  role!: Exclude<Role, "OWNER">;
  @ApiProperty({ enum: InvitationStatus })
  status!: InvitationStatus;
  invitedBy!: string;
  expiresAt!: Date;
  acceptedAt!: Date | null;
  createdAt!: Date;
}